Refactor server side
